+++ tests/framework/hooks/test_hook_log_metrics.py
@@ -157,10 +157,10 @@ def test_mlflow_hook_automatically_prefix_metrics_dataset(
         )
         # Check if metrics datasets have prefix with its names.
         # for metric
-        assert dummy_catalog._data_sets["my_metrics"]._prefix == "my_metrics"
-        assert dummy_catalog._data_sets["another_metrics"]._prefix == "foo"
-        assert dummy_catalog._data_sets["my_metric"].key == "my_metric"
-        assert dummy_catalog._data_sets["another_metric"].key == "foo"
+        assert dummy_catalog._datasets["my_metrics"]._prefix == "my_metrics"
+        assert dummy_catalog._datasets["another_metrics"]._prefix == "foo"
+        assert dummy_catalog._datasets["my_metric"].key == "my_metric"
+        assert dummy_catalog._datasets["another_metric"].key == "foo"
 
 
 def test_mlflow_hook_metrics_dataset_with_run_id(