+++ tests/mlflow/test_kedro_pipeline_model.py
@@ -263,7 +263,7 @@ def dummy_catalog(tmp_path):
 def test_model_packaging_with_copy_mode(
     tmp_path, tmp_folder, pipeline_inference_dummy, dummy_catalog, copy_mode, expected
 ):
-    dummy_catalog._data_sets["model"].save(2)  # emulate model fitting
+    dummy_catalog._datasets["model"].save(2)  # emulate model fitting
 
     kedro_model = KedroPipelineModel(
         pipeline=pipeline_inference_dummy,
@@ -293,7 +293,7 @@ def test_model_packaging_with_copy_mode(
     # second assertion: copy_mode works
     actual_copy_mode = {
         name: ds._copy_mode
-        for name, ds in loaded_model._model_impl.python_model.loaded_catalog._data_sets.items()
+        for name, ds in loaded_model._model_impl.python_model.loaded_catalog._datasets.items()
     }
 
     assert actual_copy_mode == expected
@@ -331,15 +331,15 @@ def test_model_packaging_too_many_artifacts(tmp_path, pipeline_inference_dummy):
         }
     )
 
-    catalog._data_sets["raw_data"].save(1)  # emulate input on disk
-    catalog._data_sets["model"].save(2)  # emulate model fitting
+    catalog._datasets["raw_data"].save(1)  # emulate input on disk
+    catalog._datasets["model"].save(2)  # emulate model fitting
 
     # the input is persited
     artifacts = {
         name: Path(dataset._filepath.as_posix())
         .resolve()
         .as_uri()  # weird bug when directly converting PurePosixPath to windows: it is considered as relative
-        for name, dataset in catalog._data_sets.items()
+        for name, dataset in catalog._datasets.items()
         if not isinstance(dataset, MemoryDataset)
     }
 
@@ -572,7 +572,7 @@ def test_kedro_pipeline_model_save_and_load(
         pipeline=pipeline, catalog=catalog, input_name=input_name
     )
     # emulate artifacts persistence
-    for ds in catalog._data_sets.values():
+    for ds in catalog._datasets.values():
         if hasattr(ds, "_filepath") is not None:
             ds.save(1)